DAL on getAll controller implemented
